Update support for tag prefixes
It turns out the blang/semver only supports prefixes of 'v' when
called with ParseTolerant so semverbot needs to handle remapping
or trimming that correctly before passing into semver.

This commit updates the API struct to carry the configured Prefix
and passes it down to the lower level implementation functions.

- Updated test data to include a prefix of "test-" which initially
  failed, and now passes.
- Updated test structs to include a Prefix field, used for delegation
- Added a new UpdateVersionOptions struct to pass thru to the
  UpdateVersion function, following the same structure as existing code

Fixes #58